requirements.txt
matplotlib==3.8.4
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
py_cpuinfo==9.0.0
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.14.1
skimage==0.0
timm==1.0.11
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
